Introduce injection container
