+++ pubspec.yaml
@@ -28,7 +28,6 @@ dependency_overrides:
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.8
 
   # isar dependencies
   isar: ^3.1.0+1
@@ -39,11 +38,13 @@ dependencies:
   flutter_riverpod: ^2.5.1
   riverpod_annotation: ^2.3.5
 
+  # other dependencies
+  cupertino_icons: ^1.0.8
+  get_it: ^7.7.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^4.0.0
-  build_runner: ^2.4.11
 
   # isar dependencies
   isar_generator: ^3.1.0+1
@@ -52,6 +53,10 @@ dev_dependencies:
   riverpod_generator: ^2.4.2
   custom_lint: ^0.6.4
   riverpod_lint: ^2.3.12
+
+  # other dev dependencies
+  flutter_lints: ^4.0.0
+  build_runner: ^2.4.11
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
